fix : is_ascii function moved to font_wizard.py
