+++ otherfile/art_decor_check.py
@@ -5,6 +5,7 @@ import art
 from art.art_param import DECORATIONS_MAP as Decor_Dict
 from art.art_dic import art_dic as Art_Dict
 from art.art_param import NON_ASCII_ARTS
+from font_wizard import is_utf8,is_ascii
 
 Failed1 = 0
 Failed2 = 0
@@ -14,38 +15,6 @@ Message1 = "Art UTF-8 compatibility test "
 Message2 = "Art duplication test "
 Message3 = "Decor UTF-8 compatibility test "
 Message4 = "Decor duplication test "
-
-
-def is_utf8(s):
-    """
-    Check input string for UTF-8 compatibility.
-
-    :param s: input string
-    :type s: str
-    :return: result as bool
-    """
-    try:
-        if sys.version_info.major == 3:
-            _ = bytes(s, 'utf-8').decode('utf-8', 'strict')
-        else:
-            return True
-        return True
-    except Exception:
-        return False
-
-def is_ascii(s):
-    """
-    Check input string for ASCII compatibility.
-
-    :param s: input string
-    :type s: str
-    :return: result as bool
-    """
-    for i in s:
-        if ord(i) > 127:
-            return False
-    return True
-
 
 def print_result(flag_list, message_list):
     """
